requirements.txt
selenium
pandas
openai
chromedriver-autoinstaller
beautifulsoup4
python-dotenv
